請求項目の有効/無効条件の修正
