+++ packages/kokoas-client/src/pages/projInvoiceV2/hooks/useBillingItems.ts
@@ -1,7 +1,9 @@
 import { useContractsByProjIdV2, useInvoicesB2CByProjId } from 'kokoas-client/src/hooksQuery';
 import { useTypedWatch } from './useTypedRHF';
-import { getBillingItems } from './helper/getBillingItems';
 import { getBilledItems } from './helper/getBilledItems';
+import { TInvoiceDetails } from '../schema';
+import { getAllBillingItems } from './helper/getAllBillingItems';
+import { getBillingItems } from './helper/getBillingItems';
 
 
 
@@ -9,11 +11,13 @@ export const useBillingItems = () => {
 
   const [
     projId,
+    invoiceDetails,
   ] = useTypedWatch({
     name: [
       'projId',
+      'invoiceDetails',
     ],
-  }) as [string];
+  }) as [string, TInvoiceDetails];
 
   const {
     data: contracts,
@@ -25,16 +29,19 @@ export const useBillingItems = () => {
     isFetching: isFetchingInvoice,
   } = useInvoicesB2CByProjId(projId || '');
 
-  // 請求書から請求済みの請求項目リストを作成する
+  // 既存の請求書から請求済みの請求項目リストを作成する
   const billedItems = getBilledItems({ invoices });
 
-  const billingItems = getBillingItems({
+  // 作成中の請求書から使用中の請求項目リストを作成する
+  const billingItems = getBillingItems({ invoiceDetails });
+
+  const allBillingItems = getAllBillingItems({
     contracts,
-    billedItems,
+    billedItems: billedItems.concat(billingItems),
   });
 
   return {
-    billingItems: billingItems,
+    billingItems: allBillingItems,
     isFetching: isFetchingContract || isFetchingInvoice,
   };
 };
